[BREAKING] Disallowing shifts by signed types

+++ test/compilationTests/gnosis/Utils/Math.sol
@@ -94,12 +94,12 @@ library Math {
         zpow = zpow * z / ONE;
         result += 0x9c7 * zpow / ONE;
         if (shift >= 0) {
-            if (result >> (256-shift) > 0)
-                return (2**256-1);
-            return result << shift;
+            if (result >> uint(256 - shift) > 0)
+                return (2 ** 256 - 1);
+            return result << uint(shift);
         }
         else
-            return result >> (-shift);
+            return result >> uint(-shift);
     }
 
     /// @dev Returns natural logarithm value of given x
